chore: cleanup the APIs pkg deps and keep them compact
